Fix bug with collecting more Wakers indefinitely
For multiple polls towards the same Future the primitive
is only supposed to keep the Waker from the last poll. But the
implementation here kept all of them. This is now fixed by having each
Listener have a unique ID so each time it's polled it will overwrite the
last Waker it had stored, if any.